Bug 16223: Automatically remove any borrower debarments after a payment
Some libraries debar Patrons at the end of the year for having unpaid fines,
like in Bug 15157. Currently librarians have to manually remove this type of
debarments after Patron has paid his/her fines. Add a system preference to
define debarments that should be automatically removed after a payment is made,
and add functionality to actually remove the defined debarments from Patron.
Also let libraries to define the amount of outstanding fines after payment
after which the debarment will be removed.

This patch introduces a system preference DebarmentsToLiftAfterPayment, which
allows libraries to define rules for removing debarments after paying fines.
The system preference uses YAML and is defined as follows:

Debarment with this comments will be removed:
  outstanding: 5

Which means that if a Patron has a debarment "Debarment with this comment will
be removed", and he pays his fines and charges until his outstanding fees are
equal or less than 5.00, this debarment will be lifted. The parameter outstanding
is optional - if not given, the debarment will be lifted until Patron has paid
all of his outstanding fees (in other words, equal to "outstanding: 0").

To test:
1. Set a debarment to a Patron
2. Set a fine for Patron
3. Define rule(s) for removing debarment(s) in system preference
   DebarmentsToLiftAfterPayment
4. Pay the fine you set in step 2
5. Note that the debarment is now lifted

Also prove t/db_dependent/Patron/Borrower_Debarments.t

Rebased-by: Emmi Takkinen <emmi.takkinen@koha-suomi.fi>

Signed-off-by: Katariina Hanhisalo <katariina.hanhisalo@xamk.fi>