Bug 17712: Add base for item and biblio availability
This patch adds two new classes:
- Koha::Biblio::Availability
- Koha::Item::Availability

These classes represent biblio- and item-level availabilities and the purpose
of this patch is simply to provide constructors that validate given biblio/item
and patron parameters.

Hold / checkout / search view availabilities (the actual answers to availability)
will be provided in the next patches.

To test:
1. Run t/db_dependent/Koha/Item/Availability.t
2. Run t/db_dependent/Koha/Biblio/Availability.t