KD-0: Add option to clear other cached values related to a systempreference
Some features in Koha cache other values depending on a value of a specific
system preference. Because of this, we need to be able to clear any related caches
when the value of a system preference is modified.

An example of this is parsed YAML configuration of a system preference. By caching
the parsed YAML, we can optimize our code by avoiding the need of parsing the YAML
on each call. As the system preference gets modified, we then need to be able to
clear parsed YAML from the cache.