KD-0 Plack initscript improvements - Revert log/run-dir permissions overriding and directory creation
During a process of trying to find a solution to the privilege issues of running Koha-processes,
and subsequent log/run-dir permission overhaul,
found that koha-plack-daemon forcibly overrides ownerships.

This breaks for ex stunnel4-client from being able to write to /var/log/koha/stunnel/*

This type of behaviour doesn't belong to the plack-daemon, but must be in the
Koha installer itself.