KD-3750: Fix SRU logging settings
Only upstream version has logs directory, Koha-suomi master doesn't
have it. Also creating the log during logtime makes sure the right
process/user is opening the file so then for example if the Koha
process is run as www-data it doesn't halt on this when it cannot open
the log file.