Bug 23449: Hide Biblio::PickupLocations from holdability API when not requested
When calling Koha::Biblio::Availability::Hold with query_pickup_locations
disabled, the availability notes would still include the note
Koha::Exceptions::Biblio::PickupLocations. We don't want it to be there
in this case.

This patch fixes a bug that caused it to be shown even when not requested.

To test:
1. prove t/db_dependent/Koha/Biblio/Availability/Hold.t
2. prove t/db_dependent/api/v1/availability.t
3. prove t/db_dependent/Koha/Item/Availability/Hold/Opac/ItemStatus.t