Fix for KD-543 / 8f1b7b8884 after which only half of the items became added,
for some reason, $i++ re-invoked in the 3rd section (no such in KohaCommunity btw) in the for loop, but it also was there as increment if successful adding happened.

Also, there was a reasonable warning that loop might become infinite if it loops over, let's say, the same bar code number returned always,

so at least to detect this error and prevent Perl process from hanging in the infinite loop, I've added another counter to die if it grows too much.

I also created detection that it fails on the big difference between really added and skipped items only. So it can go beyond 10_000 in that loop if almost most of those items added.