KD-543: (squashable) Add requested amount of items instead of half the amount
The i++ being in the for clause is sufficient
because the $barcodeobj->next_value and
C4::Barcodes::ValueBuilder::hbyyyyincr::get_barcode statements
return a valid barcode always except when we are out of barcodes.

Therefore, we can remove the other i++ statement inside the for a loop.
In the case where we run out of barcodes, this adds the error
"no_next_barcode" to the list of errors.

(also related: KD-4000 and 8f1b7b8)