KD-4343: Add UserJS checksumming
Add optional checksumming for IntranetUserJS and
OPACUserJS systempreferences. This will prevent malicious
changes to the preferences by stopping the Koha UI when
unauthorized changes are detected.

The checking is done only if the checksums are stored in
koha-conf, in which case you will need to update the
checksums whenever the preferences are changed by
running misc/rehash_userjs_sysprefs.pl.