KD-4351: Cronscript for resetting the failed login attempts
This cronjob will reset the amount of failed login attempts after
specified amount of time. This will lift the bans caused by the
user trying to log in with wrong password too many times.
Resetting is done based on the borrowercategory.

The cronjob will need be run at appropriate intervals depending
on the desired length of the bans. For example to reset failed
login attempts that are older than 3600 seconds every hour add
this to your crontab:

*/60 * * * * $KOHA_PATH/misc/cronjobs/reset_failed_logins.pl 3600 CATEGORYCODE

The time is calculated from the last time the user tried to log in,
so each consecutive failed login while the account is locked
will push the ban forward.