KD-4355: Change how "pull this many items" is counted on holds to pull
Previously holds to pull counted the amount of individual patrons
that have holds on a title and reported this as "pull this many
items". This is changed so that individual holds are counted
instead of patrons. This will give the staff a better idea on how
many items actually need to be pulled.