Set dry mode as default for Ropocapital print provider
This is for new installations.

Dry-mode was disabled in KOHA_CONF and if overduemessages.pl were to be
executed, it would had made a request to Ropocapital server with invalid
logins.

Additionally, remove client id from the configuration. It is not
sensitive data but it should definitely not be there.

Also removes the real hostname. Let sysadmins configure it.