Bug 12133: Fix issues with modifying child patron and adding guarantors
Child patrons couldn't be modified when guarantor was
patron due missing guarantor_id value. Also if multiple
guarantors were added one could easily add child patron
as a guarantor.

This patch adds Instead of one guarantor id all new and
existing guarantor ids are checked in case of child patron.

To test:
1. Add child patron with syspref ChildNeedsGuarantor and
GuarantorNeedsToBePatron on.
2. Add two guarantors for the patron, first one should be
an adult patron and second a child.
3. Save patron.
=> Error "A guarantor cannot be a guarantee." should be
raised but isn't.
4. Modify patron and save.
=> Error "Child needs guarantor" is raised even if (valid)
guarantor exists.
5. Apply patch.
6. Repeat steps 1 to 4.
=> This time while saving a new patron error "A guarantor
cannot be a guarantee." is raised but modifying is
successful.

Sponsored-by: Koha-Suomi Oy