KD-4838: Add --user when pidfile is used in koha-plack-daemon
Since D10 start-stop-daemon changed its behaviour. From
start-stop-daemon manual:

"Warning: using this match option with a world-writable
pidfile or using it alone with a daemon that writes the
pidfile as an unprivileged (non-root) user will be refused
with an error (since version 1.19.3) as this is a security
risk, because either any user can write to it, or if the
daemon gets compromised, the contents of the pidfile
cannot be trusted, and then a privileged runner (such as
an init script executed as root) would end up acting on
any system process. Using /dev/null is exempt from these
checks."

This is fixed in community version with bug 25481.