KD-4928 Borrowercategory upper age limit fix
Checking the upper age limit for borrowercategory when adding or
modifying borrowers doesn't work as described in Koha manual.
j2a.pl script however doe follow the manual, which causes the SC
and j2a.pl script to behave differently.

This fixed the SC memberentry.pl and update-child.pl scripts to
work as described in the manual.