K21-398 Revert to older version of translator
Koha version 21.11 uses gulp to process translations. Gulp is built
on node.js, which has no use whatsoever beyond the relatively
simple task of maintaining po-files. Since we do not want the
maintenance and security burdens caused by having to have node.js
installed just for this, we will revert to the older version of
translator.